requirements.txt
aiosqlite==0.21.0
alembic==1.16.2
anyio==4.9.0
bcrypt==4.0.1
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
ecdsa==0.19.1
fastapi==0.109.2
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
redis==6.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.36.3
typing_extensions==4.14.0
uvicorn==0.35.0
pytest
httpx
asyncpg
